requirements.txt
streamlit
graphviz
python-dotenv
authlib
rdflib
